+++ lib/cache.sh
@@ -171,14 +171,25 @@ function cache::save() {
 	output::step "Saving cache"
 
 	mkdir -p "${cache_dir}/.heroku"
-
 	rm -rf "${cache_dir}/.heroku/python"
-	# In theory we should be able to use `--reflink=auto` here for improved performance, however,
-	# initial benchmarking showed it to be slower with the file system type / mounts used by the
-	# Heroku build system for some reason. (Copying was faster using `--link`, however, that fails
-	# when copying cross-mount such as for Heroku CI and build-in-app-dir, plus hardlinks could
-	# result in unintended cache mutation if later buildpacks add/remove packages etc.)
-	cp --recursive "${build_dir}/.heroku/python" "${cache_dir}/.heroku/"
+
+	local build_dir_filesystem cache_dir_filesystem
+	build_dir_filesystem="$(df --output=target "${build_dir}")"
+	cache_dir_filesystem="$(df --output=target "${cache_dir}")"
+
+	# For improved performance, we copy using hard-links if possible. This requires that the build
+	# and cache directory are on the same filesystem mount - which is the case for standard builds
+	# but not Heroku CI or build-in-app-dir. Ideally we would be able to use `--reflink=auto` here
+	# (which would avoid the need for a conditional and also mean accidental edits by users in later
+	# buildpacks to one location doesn't affect the other), however, with the current filesystems
+	# used in production benchmarking showed `--reflinks=auto` was much slower than hardlinks.
+	if [[ "${build_dir_filesystem}" == "${cache_dir_filesystem}" ]]; then
+		local additional_copy_args=(--link)
+	else
+		local additional_copy_args=()
+	fi
+
+	cp --recursive "${additional_copy_args[@]}" "${build_dir}/.heroku/python" "${cache_dir}/.heroku/"
 
 	# Metadata used by subsequent builds to determine whether the cache can be reused.
 	# These are written/consumed via separate files and not the build data store for compatibility